ms_admin/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-mysql==4.2.0
djangorestframework==3.12.4
mysqlclient==2.0.3
pika==1.2.0
pytz==2021.3
sqlparse==0.4.2
